refactor: MeshFactory and Builder classes moved from V2 one level up (as there is no v1 for months).
