+++ Assets/UnZENity-Core/Scripts/Caches/StaticCache/VobCacheCreator.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
-using GUZ.Core.Creator.Meshes.V2;
+using GUZ.Core.Creator.Meshes;
 using GUZ.Core.Globals;
 using GUZ.Core.Vm;
 using UnityEngine;
