+++ Assets/UnZENity-Core/Scripts/Creator/Meshes/Builder/NpcMeshBuilder.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 using ZenKit;
 using Vector3 = System.Numerics.Vector3;
 
-namespace GUZ.Core.Creator.Meshes.V2.Builder
+namespace GUZ.Core.Creator.Meshes.Builder
 {
     public class NpcMeshBuilder : AbstractMeshBuilder
     {
