+++ Assets/UnZENity-Core/Scripts/Creator/Meshes/Builder/VobMeshBuilder.cs
@@ -2,7 +2,7 @@ using GUZ.Core.Extensions;
 using GUZ.Core.Globals;
 using UnityEngine;
 
-namespace GUZ.Core.Creator.Meshes.V2.Builder
+namespace GUZ.Core.Creator.Meshes.Builder
 {
     public class VobMeshBuilder : AbstractMeshBuilder
     {
