+++ Assets/UnZENity-Core/Scripts/Creator/Meshes/MeshFactory.cs
@@ -1,6 +1,6 @@
 using System.Linq;
 using System.Threading.Tasks;
-using GUZ.Core.Creator.Meshes.V2.Builder;
+using GUZ.Core.Creator.Meshes.Builder;
 using GUZ.Core.Data.Container;
 using GUZ.Core.Extensions;
 using GUZ.Core.Manager;
@@ -10,7 +10,7 @@ using ZenKit;
 using ZenKit.Daedalus;
 using ZenKit.Vobs;
 
-namespace GUZ.Core.Creator.Meshes.V2
+namespace GUZ.Core.Creator.Meshes
 {
     public static class MeshFactory
     {
