--- Assets/UnZENity-Core/Scripts/Creator/Meshes/V2.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 6359f4eae4c5403082546ee247e3752a
-timeCreated: 1710789860
