+++ Assets/UnZENity-Lab/Scripts/Handler/LabVobItemHandler.cs
@@ -3,7 +3,7 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using GUZ.Core;
-using GUZ.Core.Creator.Meshes.V2;
+using GUZ.Core.Creator.Meshes;
 using GUZ.Core.Globals;
 using GUZ.Core.Properties;
 using GUZ.Core.Vm;
